@@ -1,8 +1,11 @@
 import { Component, HostBinding, HostListener, OnInit, OnDestroy } from '@angular/core';
-import { Router } from '@angular/router'
+import { Router } from '@angular/router';
+import { FormGroup, FormControl, Validators } from "@angular/forms";
+
 
 import { AppService } from '../app.service';
 import { Subscription } from 'rxjs/Subscription';
+
 
 import { backgroundTrigger, overlayTrigger, contentTrigger } from '../animations';
 
@@ -13,6 +16,7 @@ import { backgroundTrigger, overlayTrigger, contentTrigger } from '../animations
   animations: [ backgroundTrigger, overlayTrigger, contentTrigger ]
 })
 export class ContactComponent implements OnInit {
+myForm: FormGroup;
 subscription: Subscription;
 backgroundState: string;
 overlayState: string;
@@ -26,6 +30,12 @@ route: string;
     this.backgroundState = 'show';
     this.overlayState = 'hidden';
     this.contentState = 'hidden';
+    this.myForm = new FormGroup({
+      name: new FormControl(null, Validators.required),
+      email: new FormControl(null, Validators.required),
+      subject: new FormControl(null, Validators.required),
+      body: new FormControl(null, Validators.required)
+    })
     this.subscription = this.appService.navigationChanged
       .subscribe(
         (route: string) => {
@@ -51,6 +61,20 @@ route: string;
         this.router.navigateByUrl('/' + this.route);
     }
   }
+
+
+    submitContactForm(){
+        let emailBundle = {
+          name: this.myForm.value.name,
+          email: this.myForm.value.email,
+          subject: this.myForm.value.subject,
+          body: this.myForm.value.body
+        }
+
+        this.appService.submitContact(emailBundle);
+        this.myForm.reset();
+
+    }
 
   ngOnDestroy() {
     this.subscription.unsubscribe();
